Work on part 2

+++ src/main/java/aoc/y2019/day22/CutCards.java
@@ -1,30 +1,30 @@
 package aoc.y2019.day22;
 
 public class CutCards implements Technique {
-    private int numCards;
+    private long numCards;
 
-    public CutCards(int numCards) {
+    public CutCards(long numCards) {
         this.numCards = numCards;
     }
 
     @Override
-    public int apply(int size, int index) {
+    public long apply(long size, long index) {
         return numCards < 0 ? cutEnd(size, index) : cutFront(size, index);
     }
 
-    private int cutEnd(int size, int index) {
+    private long cutEnd(long size, long index) {
         var cut = size - Math.abs(numCards) - 1;
 
         return doCut(size, index, cut);
     }
 
-    private int cutFront(int size, int index) {
+    private long cutFront(long size, long index) {
         var cut = numCards - 1;
 
         return doCut(size, index, cut);
     }
 
-    private int doCut(int size, int index, int cut) {
+    private long doCut(long size, long index, long cut) {
         if (index <= cut) {
             var endNdx = size - 1;
             var diff = cut - index;
