+++ src/main/java/aoc/y2019/day22/DealStack.java
@@ -2,7 +2,7 @@ package aoc.y2019.day22;
 
 public class DealStack implements Technique {
     @Override
-    public int apply(int size, int index) {
+    public long apply(long size, long index) {
         var endNdx = size - 1;
         return endNdx - index;
     }
