+++ src/main/java/aoc/y2019/day22/Part1.java
@@ -2,17 +2,17 @@ package aoc.y2019.day22;
 
 import aoc.utils.Problem;
 
-public class Part1 extends Problem<Integer> {
+public class Part1 extends Problem<Long> {
     private Parser parser;
 
-    public Part1(String fileName, int exp) {
+    public Part1(String fileName, long exp) {
         super(exp);
 
         parser = new Parser(fileName);
     }
 
     @Override
-    public Integer run() {
+    public Long run() {
         var shuffler = parser.parse();
 
         return shuffler.apply(10007, 2019);
