+++ src/main/java/aoc/y2019/day22/Technique.java
@@ -1,5 +1,5 @@
 package aoc.y2019.day22;
 
 public interface Technique {
-    int apply(int size, int index);
+    long apply(long size, long index);
 }
